+++ packages/liveblocks-core/src/crdts/LiveObject.ts
@@ -14,13 +14,13 @@ import type {
 } from "../protocol/Op";
 import { OpCode } from "../protocol/Op";
 import type {
-  IdTuple,
-  SerializedChild,
-  SerializedCrdt,
+  NodeStream,
+  ObjectStorageNode,
+  RootStorageNode,
   SerializedObject,
   SerializedRootObject,
-} from "../protocol/SerializedCrdt";
-import { CrdtType } from "../protocol/SerializedCrdt";
+} from "../protocol/StorageNode";
+import { CrdtType, isRootStorageNode } from "../protocol/StorageNode";
 import type * as DevTools from "../types/DevToolsTreeNode";
 import type { ParentToChildNodeMap } from "../types/NodeMap";
 import type { ApplyResult, ManagedPool } from "./AbstractCrdt";
@@ -52,10 +52,6 @@ export type LiveObjectUpdates<TData extends LsonObject> = {
   node: LiveObject<TData>;
   updates: LiveObjectUpdateDelta<TData>;
 };
-
-function isRootCrdt(id: string, _: SerializedCrdt): _ is SerializedRootObject {
-  return id === "root";
-}
 
 /**
  * The LiveObject class is similar to a JavaScript object that is synchronized on all clients.
@@ -92,21 +88,21 @@ export class LiveObject<O extends LsonObject> extends AbstractCrdt {
   public static detectLargeObjects = false;
 
   static #buildRootAndParentToChildren(
-    items: IdTuple<SerializedCrdt>[]
+    nodes: NodeStream
   ): [root: SerializedRootObject, nodeMap: ParentToChildNodeMap] {
     const parentToChildren: ParentToChildNodeMap = new Map();
     let root: SerializedRootObject | null = null;
 
-    for (const [id, crdt] of items) {
-      if (isRootCrdt(id, crdt)) {
-        root = crdt;
+    for (const node of nodes) {
+      if (isRootStorageNode(node)) {
+        root = node[1];
       } else {
-        const tuple: IdTuple<SerializedChild> = [id, crdt];
+        const crdt = node[1];
         const children = parentToChildren.get(crdt.parentId);
         if (children !== undefined) {
-          children.push(tuple);
+          children.push(node);
         } else {
-          parentToChildren.set(crdt.parentId, [tuple]);
+          parentToChildren.set(crdt.parentId, [node]);
         }
       }
     }
@@ -120,11 +116,11 @@ export class LiveObject<O extends LsonObject> extends AbstractCrdt {
 
   /** @private Do not use this API directly */
   static _fromItems<O extends LsonObject>(
-    items: IdTuple<SerializedCrdt>[],
+    nodes: NodeStream,
     pool: ManagedPool
   ): LiveObject<O> {
     const [root, parentToChildren] =
-      LiveObject.#buildRootAndParentToChildren(items);
+      LiveObject.#buildRootAndParentToChildren(nodes);
     return LiveObject._deserialize(
       ["root", root],
       parentToChildren,
@@ -178,7 +174,7 @@ export class LiveObject<O extends LsonObject> extends AbstractCrdt {
 
   /** @internal */
   static _deserialize(
-    [id, item]: IdTuple<SerializedObject | SerializedRootObject>,
+    [id, item]: RootStorageNode | ObjectStorageNode,
     parentToChildren: ParentToChildNodeMap,
     pool: ManagedPool
   ): LiveObject<LsonObject> {
@@ -198,8 +194,9 @@ export class LiveObject<O extends LsonObject> extends AbstractCrdt {
       return liveObj;
     }
 
-    for (const [id, crdt] of children) {
-      const child = deserializeToLson([id, crdt], parentToChildren, pool);
+    for (const node of children) {
+      const child = deserializeToLson(node, parentToChildren, pool);
+      const crdt = node[1];
       if (isLiveStructure(child)) {
         child._setParentLink(liveObj, crdt.parentKey);
       }